Remove legacy verses test file
